make util.js file and change all size to rem
