+++ css/intro.css
@@ -7,7 +7,7 @@
 }
 
 .loading {
-  margin: 39vh 0;
+  margin: 23rem auto;
   text-align: center;
 }
 
@@ -29,20 +29,20 @@
 
 .toast-box {
   position: absolute;
-  top: 30px;
-  right: 30px;
+  top: 2rem;
+  right: 2rem;
   display: flex;
   align-items: flex-end;
   overflow: hidden;
-  padding: 20px;
+  padding: 1rem;
 }
 
 .toast {
-  width: 320px;
-  height: 55px;
+  width: 20rem;
+  height: 3.5rem;
   background: #fff;
   font-weight: 500;
-  margin: 15px 0;
+  margin: 1rem 0;
   box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
   display: flex;
   align-items: center;
@@ -50,8 +50,8 @@
 }
 
 .toast i {
-  margin: 0 20px;
-  font-size: 35px;
+  margin: 0 1rem;
+  font-size: 2rem;
   color: red;
 }
 
@@ -61,7 +61,7 @@
   left: 0;
   bottom: 0;
   width: 100%;
-  height: 5px;
+  height: 0.3rem;
   background: red;
   animation: anim 5s linear forwards;
 }
@@ -77,12 +77,6 @@
 }
 
 @media screen and (max-width: 500px) {
-  .user-input {
-    font: 300 17px 'Roboto';
-    width: 22rem;
-    line-height: 32px;
-  }
-
   .not-found-text,
   .not-authorized-text {
     font-size: 1rem;
